move controller_execute to separate file
This prevents loading of OpenMP for standalone simulator which shouldn't require the fix for double initialization.
